# Conflicts:
#	rqt_bag/src/rqt_bag/plugins/raw_view.py
#	rqt_bag_plugins/src/rqt_bag_plugins/plot_view.py

+++ rqt_bag/src/rqt_bag/timeline_frame.py
@@ -1005,11 +1005,11 @@ class TimelineFrame(QGraphicsItem):
 
         if clamp_to_visible:
             if fraction <= 0.0:
-                return int(self._stamp_left)
+                return self._stamp_left
             elif fraction >= 1.0:
-                return int(self._stamp_right)
+                return self._stamp_right
 
-        return int(self._stamp_left + fraction * (self._stamp_right - self._stamp_left))
+        return self._stamp_left + fraction * (self._stamp_right - self._stamp_left)
 
     def map_dx_to_dstamp(self, dx):
         """
